requirements.txt
Flask==1.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
loguru==0.3.2
psycopg2==2.7.7
PyYAML==5.1
python-dateutil==2.8.0
numpy==1.17.3
requests==2.21.0
tqdm==4.39.0
